refactor(pr07): delete settings/constants and settings/_types, move composites and types to shared/ (G5f.7)
Move SETTINGS/SETTING_MENUS/choices/defaults/constants composites + ConstantTypes from settings/constants/index.mts to settings/shared/constants.mts. Move SettingScope/SettingConfig<T>/SystemSettings from settings/_types.mts to settings/shared/types.mts. Delete the two aggregator files.

shared/index.mts re-exports both modules; settings/index.mts re-exports shared via export-star, preserving @settings/* back-compat. helpers.mts now imports SETTINGS from shared/constants.mjs, keeping the generic <K extends keyof typeof SETTINGS> typing on getSetting/setSetting (no behavior change).
