+++ src/settings/index.mts
@@ -1,13 +1,16 @@
 /**
  * System Settings Module
  *
- * This module provides access to settings constants, types, and utilities.
+ * Provides access to settings constants, types, and utilities.
  * Organized by category following the entity folder pattern.
  */
 
-// Export types
-export type * from './_types.mjs';
-export type { ConstantTypes } from './constants/index.mjs';
+import { registerCombatSettings } from './combat/index.mjs';
+import { registerCoreSettings } from './core/index.mjs';
+import { registerCurrencySettings } from './currency/index.mjs';
+import { registerDisplaySettings } from './display/index.mjs';
+import { registerGameRulesSettings } from './gameRules/index.mjs';
+import { SYSTEM_ID } from './shared.mjs';
 
 // Shared
 export { SYSTEM_ID } from './shared.mjs';
@@ -23,31 +26,8 @@ export * from './roll/index.mjs';
 export * from './shared/index.mjs';
 export * from './skills/index.mjs';
 
-// Legacy exports (backwards compatibility)
-export { constants } from './constants/index.mjs';
-export { helpers } from './helpers.mjs';
-
-// Re-export for convenience (backwards compatibility)
-export {
-  DEFAULT_CURRENCY_CONFIG,
-  DEFAULT_HEALTH_CONFIG,
-  DEFAULT_ROLL_CONFIG,
-  DIAGONAL_MOVEMENT_CHOICES,
-  EXPERIENCE_RATE_CHOICES,
-  PARTY_HUD_CHOICES,
-  SETTING_MENUS,
-  SETTINGS,
-  SHARED_VISION_MODE_CHOICES,
-  UNIT_CHOICES,
-} from './constants/index.mjs';
-export { getSetting, setSetting } from './helpers.mjs';
-
-import { registerCombatSettings } from './combat/index.mjs';
-import { registerCoreSettings } from './core/index.mjs';
-import { registerCurrencySettings } from './currency/index.mjs';
-import { registerDisplaySettings } from './display/index.mjs';
-import { registerGameRulesSettings } from './gameRules/index.mjs';
-import { SYSTEM_ID as SYSTEM_ID_INTERNAL } from './shared.mjs';
+// Helpers
+export { getSetting, helpers, setSetting } from './helpers.mjs';
 
 /**
  * Register all system settings.
@@ -57,7 +37,7 @@ import { SYSTEM_ID as SYSTEM_ID_INTERNAL } from './shared.mjs';
  * its menu button(s) and its individual settings.
  */
 function registerSettings(): void {
-  console.log(`${SYSTEM_ID_INTERNAL} | Registering system settings`);
+  console.log(`${SYSTEM_ID} | Registering system settings`);
 
   registerCoreSettings();
   registerGameRulesSettings();
@@ -65,8 +45,7 @@ function registerSettings(): void {
   registerDisplaySettings();
   registerCurrencySettings();
 
-  console.log(`${SYSTEM_ID_INTERNAL} | Settings registration complete`);
+  console.log(`${SYSTEM_ID} | Settings registration complete`);
 }
 
 export { registerSettings };
-
