+++ src/settings/shared/index.mts
@@ -1,9 +1,10 @@
 /**
  * Shared settings infrastructure
  *
- * Cross-cutting helpers, components, and stores used by multiple settings
- * categories (and elsewhere in the system). Currently re-exports the generic
- * settings sheet plumbing; intended to grow as more shared concerns emerge.
+ * Cross-cutting helpers, components, stores, composites, and types used by
+ * multiple settings categories (and elsewhere in the system).
  */
 
+export * from './constants.mjs';
 export * from './sheet/index.mjs';
+export type { SettingConfig, SettingScope, SystemSettings } from './types.mjs';
